backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
passlib>=1.7.4
pydantic>=2.7.0
aiosqlite>=0.20.0
